refactor / move plugin code in a separate folder, use plugin theme
