+++ hat/assets/js/apps/Iaso/domains/app/components/LocalizedAppComponent.js
@@ -1,8 +1,8 @@
 /* eslint-disable import/no-relative-packages */
+import React, { useContext } from 'react';
 import { AdapterMoment } from '@mui/x-date-pickers/AdapterMoment';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import PropTypes from 'prop-types';
-import React, { useContext } from 'react';
 import { IntlProvider } from 'react-intl';
 
 // the intl paths get rewritten by webpack depending on the locale
@@ -11,7 +11,7 @@ import fr from '__intl/messages/fr'; // eslint-disable-line
 
 import enLibrary from '../../../../../../../../node_modules/bluesquare-components/dist/locale/en.json';
 import frLibrary from '../../../../../../../../node_modules/bluesquare-components/dist/locale/fr.json';
-import { PluginsContext } from '../../../utils/index.ts';
+import { PluginsContext } from '../../../plugins/context';
 import { useLocale } from '../contexts/LocaleContext.tsx';
 
 const extractTranslations = (plugins, key) => {
