+++ hat/assets/js/apps/Iaso/domains/app/hooks/useRoutes.tsx
@@ -1,5 +1,4 @@
-import { last } from 'lodash';
-import React, { ElementType, ReactElement, useContext, useMemo } from 'react';
+import React, { ReactElement, useMemo } from 'react';
 import { Route, Routes } from 'react-router';
 
 import {
@@ -9,50 +8,22 @@ import {
 } from '../../../constants/routes';
 
 import { baseUrls } from '../../../constants/urls';
+import {
+    useHomeOnlineComponent,
+    useHomeOfflineComponent,
+    usePluginsRouteConfigs,
+} from '../../../plugins/hooks/routes';
 import { useRedirections } from '../../../routing/hooks/useRedirections';
 import { RouteCustom } from '../../../routing/types';
-import { PluginsContext } from '../../../utils';
-import { SHOW_HOME_ONLINE, hasFeatureFlag } from '../../../utils/featureFlags';
 import { useCurrentUser, useHasNoAccount } from '../../../utils/usersUtils';
-import { HomeOnline } from '../../home/HomeOnline';
 import ProtectedRoute from '../../users/components/ProtectedRoute';
 import { useGetCurrentUser } from '../../users/hooks/useGetCurrentUser';
-import { Plugins } from '../types';
 
 type Result = {
     routes: ReactElement | null;
     nonDashboardRoutes: ReactElement | null;
     isLoadingRoutes: boolean;
     isCurrentRouteAnonymous: boolean;
-};
-
-const useHomeOnlineComponent = (): ElementType | undefined => {
-    const { plugins }: Plugins = useContext(PluginsContext);
-    const currentUser = useCurrentUser();
-    const canShowHome = hasFeatureFlag(currentUser, SHOW_HOME_ONLINE);
-    const PluginHome = last(
-        plugins
-            .filter(plugin => plugin.homeOnline)
-            .map(plugin => plugin.homeOnline),
-    );
-    // using the last plugin override (arbitrary choice)
-    return useMemo(
-        () => (canShowHome ? PluginHome || HomeOnline : undefined),
-        [PluginHome, canShowHome],
-    );
-};
-export const useHomeOfflineComponent = (): ElementType | undefined => {
-    const { plugins }: Plugins = useContext(PluginsContext);
-    // using the last plugin override (arbitrary choice)
-    return useMemo(
-        () =>
-            last(
-                plugins
-                    .filter(plugin => plugin.homeOffline)
-                    .map(plugin => plugin.homeOffline),
-            ),
-        [plugins],
-    );
 };
 
 const useHomeOnlineRoute = (userHomePage?: string): RouteCustom[] => {
@@ -85,21 +56,6 @@ export const useHomeOfflineRoute = (): RouteCustom[] => {
             element: <HomeComponent />,
         },
     ];
-};
-
-const usePluginsRouteConfigs = (): {
-    pluginRoutes: RouteCustom[];
-    pluginRedirections: any[];
-} => {
-    const { plugins }: Plugins = useContext(PluginsContext);
-    const pluginRoutes = plugins.map(plugin => plugin.routes).flat();
-    const pluginRedirections = plugins
-        .map(plugin => plugin.redirections)
-        .flat();
-    return useMemo(
-        () => ({ pluginRoutes, pluginRedirections }),
-        [pluginRedirections, pluginRoutes],
-    );
 };
 
 const useGetProtectedRoutes = (
