+++ hat/assets/js/apps/Iaso/domains/app/types.ts
@@ -1,10 +1,5 @@
-import { ElementType, ReactNode } from 'react';
-import { Theme } from '@mui/material/styles';
+import { ReactNode } from 'react';
 import { IntlMessage } from 'bluesquare-components';
-import {
-    RouteCustom,
-    Redirection as RoutingRedirection,
-} from '../../routing/types';
 
 export type MenuItem = {
     label: string | IntlMessage;
@@ -24,28 +19,6 @@ export type MenuItems = MenuItem[];
 export type Redirection = {
     path: string;
     component: (args: any) => ReactNode;
-};
-
-export type Plugin = {
-    routes: RouteCustom[];
-    menu: MenuItem[];
-    translations: Record<string, any>;
-    homeUrl?: string;
-    homeOnline?: ElementType;
-    homeOffline?: ElementType;
-    key?: string;
-    baseUrls: Record<string, string>;
-    paramsConfig: Record<string, string[]>;
-    redirections?: RoutingRedirection[];
-    customComponents?: {
-        key: string;
-        component: ElementType;
-    }[];
-    theme?: Theme;
-};
-
-export type Plugins = {
-    plugins: Plugin[];
 };
 
 export type PaginatedResponse<T> = {
