+++ hat/assets/js/apps/Iaso/plugins/context.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+import { Plugin } from './types';
+
+export const PluginsContext = createContext<{ plugins: Plugin[] }>({
+    plugins: [],
+});
