+++ hat/assets/js/apps/Iaso/routing/hooks/useRedirections.tsx
@@ -1,11 +1,11 @@
-import { getSort } from 'bluesquare-components';
 import React, { ReactElement } from 'react';
+import { getSort } from 'bluesquare-components';
 import { Route } from 'react-router-dom';
 import Page404 from '../../components/errors/Page404';
 import { baseUrls } from '../../constants/urls';
-import { useHomeOfflineComponent } from '../../domains/app/hooks/useRoutes';
 import { defaultSorted as storageDefaultSort } from '../../domains/storages/config';
 import { defaultSorted as workflowDefaultSort } from '../../domains/workflows/config/index';
+import { useHomeOfflineComponent } from '../../plugins/hooks/routes';
 import { SHOW_HOME_ONLINE, hasFeatureFlag } from '../../utils/featureFlags';
 import { useCurrentUser } from '../../utils/usersUtils';
 import { Redirect } from '../Redirect';
